requirements.txt
requests==2.31.0
python-dotenv==1.0.0
openai==1.11.1
colorama==0.4.6
rich==13.6.0
typer==0.9.0 